Day 12, rolled out of bed for this
After a few days of learning to just do what you're told, a puzzle
where you absolutely mustn't. Running the script was expensive so
the off-by-1 between path and hist, and my decision to change my
metric between parts 1 and 2, cost me a ton of time. Hadd to
benchmark the part_2 algorithm with the part_1 data to save things.