Day 16, a long day of singing and respiratory paranoia
Part 1 was okay, but excessively literal. When it came time to
build on it for part 2 I ended up with possibilities:
- running both characters in parallel, in space-squared
- running both characters in sequence, in time-squared
With some help from Reddit I remembered how to do this recursively,
and without even calculating the targets and times thereto (which
I'd had to do to save time with my looping yet flat solutions) I
had a new part 2 written, benchmarked to the part 1 solution, and
finished running all during the runtime of my original attempt.